+++ packages/amplify-ui-components/src/index.ts
@@ -8,7 +8,6 @@ export * from './components';
 export * from './components/amplify-auth-fields/amplify-auth-fields-interface';
 export * from './components/amplify-forgot-password/amplify-forgot-password-interface';
 export * from './components/amplify-sign-in/amplify-sign-in-interface';
-export * from './components/amplify-sign-up/amplify-sign-up-interface';
 export * from './components/amplify-country-dial-code/amplify-country-dial-code-interface';
 export * from './components/amplify-select/amplify-select-interface';
 export * from './components/amplify-totp-setup/amplify-totp-setup-interface';